Update to PHP7.4, ElasticSearch 7.6.2, Ubuntu 18.04 (#353)
This PR updates most of our infrastructure to more recent versions.

PHP 7.0 -> 7.4
ElasticSearch 5.5.3 -> 7.6.2
Ubuntu 16.04LTS -> 18.04 LTS
It also:

removes the type parameter from ElasticSearch, which no longer exists
updates a lot of dependencies to newer minor versions for compatibility with PHP7.4
decreases unnecessarily high memory used by ElasticSearch from 2G to 1G
fixes numerours deprecations
properly uses booleans in DQL instead of numbers, to fix compatibility with PostgreSQL (which will be used for pull request previews on Heroku in an upcoming PR)
introduces a patch needed for the EasyAdmin bundle to work with PHP 7.4. We cannot upgrade to the next major version of the bundle until we upgrade Symfony to the next major version
If you want to try this out, you'll have to vagrant destroy your dev environment and re-create it. Gitpod should work fine, if you use the review button below.

+++ src/AppBundle/Command/AppElasticsearchReindexCommand.php
@@ -6,7 +6,6 @@ use AppBundle\Entity\Adventure;
 use AppBundle\Listener\SearchIndexUpdater;
 use AppBundle\Service\ElasticSearch;
 use Doctrine\ORM\EntityManagerInterface;
-use Elasticsearch\Common\Exceptions\Missing404Exception;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -41,15 +40,14 @@ class AppElasticsearchReindexCommand extends Command
     {
         $this
             ->setName('app:elasticsearch:reindex')
-            ->setDescription('Reindex the whole Elasticsearch index')
+            ->setDescription('Drop, create and reindex the whole Elasticsearch index')
         ;
     }
 
-    // TODO: Add when using PHP 7.1
-    /* private */ const FIELD_NON_SEARCHABLE = [
+    private const FIELD_NON_SEARCHABLE = [
         'enabled' => false,
     ];
-    const FIELD_STRING = [
+    private const FIELD_STRING = [
         'type' => 'text',
         'fields' => [
             'keyword' => [
@@ -58,13 +56,13 @@ class AppElasticsearchReindexCommand extends Command
             ],
         ],
     ];
-    const FIELD_TEXT = [
+    private const FIELD_TEXT = [
         'type' => 'text',
     ];
-    const FIELD_INTEGER = [
+    private const FIELD_INTEGER = [
         'type' => 'integer',
     ];
-    const FIELD_BOOLEAN = [
+    private const FIELD_BOOLEAN = [
         'type' => 'boolean',
     ];
     const FIELD_DATE = [
@@ -75,16 +73,13 @@ class AppElasticsearchReindexCommand extends Command
     {
         $client = $this->elasticSearch->getClient();
         $indexName = $this->elasticSearch->getIndexName();
-        $typeName = $this->elasticSearch->getTypeName();
 
-        try {
+        if ($client->indices()->exists(['index' => $indexName])) {
             $client->indices()->delete([
                 'index' => $indexName,
             ]);
             $output->writeln('Deleted index.');
-        } catch (Missing404Exception $e) {
         }
-
         $client->indices()->create([
             'index' => $indexName,
         ]);
@@ -124,11 +119,8 @@ class AppElasticsearchReindexCommand extends Command
 
         $client->indices()->putMapping([
             'index' => $indexName,
-            'type' => $typeName,
             'body' => [
-                SearchIndexUpdater::TYPE => [
-                    'properties' => $mappings,
-                ],
+                'properties' => $mappings,
             ],
         ]);
         $output->writeln('Created mappings');
